+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+# [0.7.3] - 2025-07-13
+
+### Fixed
 - Improve Options#parse with glob braces support, strip quotes properly from .undercover config files
 - Fix Result#coverage_f to support ignored branches
 - Fix error parsing JSON coverage with branch coverage disabled ([#231](https://github.com/grodowski/undercover/issues/231))
@@ -177,7 +180,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - First release of `undercover` 🎉
 
-[Unreleased]: https://github.com/grodowski/undercover/compare/v0.7.2...HEAD
+[Unreleased]: https://github.com/grodowski/undercover/compare/v0.7.3...HEAD
+[0.7.3]: https://github.com/grodowski/undercover/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/grodowski/undercover/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/grodowski/undercover/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/grodowski/undercover/compare/v0.6.6...v0.7.0